+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add quick search to SFTP explorer panes.
 - Add keys support to SFTP explorer panes to navigate and perform actions over files and directories.
 - Add edit SFTP path presets from settings page.
+- Add customizable identification color for all tabs of a device.
 - Add e2e tests for SFTP UI.
 - Add Clean Tests to makefile and to code-workspace (VSCode UI - bottom bar).